+++ src/RcppExports.cpp
@@ -5,6 +5,17 @@
 
 using namespace Rcpp;
 
+// ClusterTable_new
+RcppExport SEXP ClusterTable_new(List phylo);
+RcppExport SEXP _TreeDist_ClusterTable_new(SEXP phyloSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type phylo(phyloSEXP);
+    rcpp_result_gen = Rcpp::wrap(ClusterTable_new(phylo));
+    return rcpp_result_gen;
+END_RCPP
+}
 // COMCLUST
 int COMCLUST(List trees);
 RcppExport SEXP _TreeDist_COMCLUST(SEXP treesSEXP) {