+++ tests/testthat/test-ClusterTable.R
@@ -1,0 +1,8 @@
+context("ClusterTable.R")
+library("TreeTools", quietly = TRUE)
+
+test_that("ClusterTable class behaves", {
+  tree <- BalancedTree(6)
+  as.ClusterTable(tree)
+  
+})